+++ src/i18n-l10n-editor/models/ije-folder.ts
@@ -1,4 +1,5 @@
 export interface IJEFolder {
     name: string;
     path: string;
+    arb: string;
 }