+++ src/i18n-l10n-editor/providers/ije-editor-provider.ts
@@ -8,8 +8,8 @@ export class IJEEditorProvider {
     public static register(context: vscode.ExtensionContext): vscode.Disposable {
         vscode.commands.executeCommand('setContext', 'ext:supportedFolders', IJEConfiguration.SUPPORTED_FOLDERS);
 
-        return vscode.commands.registerCommand('i18n-json-editor', (uri: vscode.Uri) => {
-            const panel = vscode.window.createWebviewPanel('i18n-json-editor', 'i18n-json-editor', vscode.ViewColumn.One, {
+        return vscode.commands.registerCommand('i18n-l10n-editor', (uri: vscode.Uri) => {
+            const panel = vscode.window.createWebviewPanel('i18n-l10n-editor', 'i18n-l10n-editor', vscode.ViewColumn.One, {
                 retainContextWhenHidden: true,
                 enableScripts: true,
                 localResourceRoots: [vscode.Uri.file(_path.join(context.extensionPath, 'media'))]