+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
-All notable changes to the "i18n-json-editor"  and "i18n-arb-editor" extension will be documented in this file.
+All notable changes to the "i18n-l10n-editor"  and "i18n-arb-editor" extension will be documented in this file.
+
+## [3.0.2]
+
+- Fixed issues with finding Workspace l10n folders
+- Fixed issue with opening editor from Status Bar
+  
+## [3.0.0]
+
+- New Version
+- Update Developer Details
+- Updated Name
+- Updated Menu Items
 
 ## [2.1.4]